getting events from calendar working
