+++ lib/tabs/events.dart
@@ -1,4 +1,3 @@
-import 'package:HackRU/tabs/sat.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:HackRU/colors.dart';
@@ -52,7 +51,7 @@ class EventsState extends State<Events>
             automaticallyImplyLeading: false,
         ),
         body: getTabBarView(<Widget>[
-          new SatEvents2(), new SunEvents()])
+          new SatEvents(), new SunEvents()])
     );
   }
 }
